CompleteTheStash: Disable hooks by default to dramatically increase the performance
You can uncomment the hooks in plugin YAML if you want missing scenes to be updated whenever you update scenes in the main Stash. This will cause each scene update to a couple of seconds which is a couple of orders of magnitudes longer than without these. The slowness is caused because each scene update will start the plugin separately and it will then connect to multiple stashes.