If a local Stash does not have a performer associated, it is still destroyed from missing Stash
This could happen if a scene is matched to Stashbox but for some reason the performer is not linked to the scene at the same time. Most of the operations to missing Stash instance is done per-performer which requires these to be deleted separately after the performers have been processed.

Fixes #3.